+++ lib/src/selectable/selectable_column_cell_wrapper.dart
@@ -1,4 +1,5 @@
-import 'package:flexible_scrollable_table_view/src/flexible_table_controller.dart';
+import 'package:flexible_scrollable_table_view/src/custom/flexible_table_controller.dart';
+import 'package:flexible_scrollable_table_view/src/flexible_table_data_source.dart';
 import 'package:flutter/widgets.dart';
 
 class SelectableColumnCellWrapper<T> extends StatelessWidget {